fix: rework profile
